Code refactor to accommodate PR comments Pinot codestyle corrections Moving ServerSegmentMetadataReader to util
