+++ pinot-server/src/main/java/org/apache/pinot/server/api/resources/TablesResource.java
@@ -150,7 +150,7 @@ public class TablesResource {
     } catch (Exception e) {
       LOGGER.error("Failed to convert table {} segment {} to json", tableName, segmentName);
       throw new WebApplicationException("Failed to convert segment metadata to json",
-              Response.Status.INTERNAL_SERVER_ERROR);
+          Response.Status.INTERNAL_SERVER_ERROR);
     } finally {
       tableDataManager.releaseSegment(segmentDataManager);
     }
@@ -248,13 +248,13 @@ public class TablesResource {
   @ApiOperation(value = "Metadata from server segment metadata.properties", notes = "Metadata from server that hosts the segment provided.")
   @ApiResponses(value = {@ApiResponse(code = 200, message = "Success"), @ApiResponse(code = 500, message = "Internal server error"), @ApiResponse(code = 404, message = "Table not found")})
   public String getSegmentReloadStatus(
-          @ApiParam(value = "Table name including type", required = true, example = "myTable_OFFLINE") @PathParam("tableName") String tableName,
-          @ApiParam(value = "Segment name", required = true) @PathParam("segmentName") String segmentName) {
+      @ApiParam(value = "Table name including type", required = true, example = "myTable_OFFLINE") @PathParam("tableName") String tableName,
+      @ApiParam(value = "Segment name", required = true) @PathParam("segmentName") String segmentName) {
     TableDataManager tableDataManager = checkGetTableDataManager(tableName);
     SegmentDataManager segmentDataManager = tableDataManager.acquireSegment(segmentName);
     if (Objects.isNull(segmentDataManager)) {
       throw new WebApplicationException(String.format("Table %s segments %s does not exist", tableName, segmentName),
-              Response.Status.NOT_FOUND);
+          Response.Status.NOT_FOUND);
     }
 
     try {
@@ -264,7 +264,7 @@ public class TablesResource {
     } catch (Exception e) {
       LOGGER.error("Failed to convert table {} segment {} to json", tableName, segmentName);
       throw new WebApplicationException("Failed to convert segment metadata to json",
-              Response.Status.INTERNAL_SERVER_ERROR);
+          Response.Status.INTERNAL_SERVER_ERROR);
     } finally {
       tableDataManager.releaseSegment(segmentDataManager);
     }
